refactor: move MultiplayerGameLayout to src/layouts/
Layout components belong in src/layouts/, not src/components/.
Update PathFight import path accordingly.
